Simplify navigation item APIs
